# swagger.yaml
openapi: 3.0.0
servers:
  - description: Localhost dev server
    url: localhost:5000
info:
  description: This is a simple API for full stack testing
  version: "1.0.0"
  title: Simple Cryptocurrency App
  contact:
    email: javierandresduarte@gmail.com
...
paths:
...
  /api/auth/signup:
    post:
      description: To register one user and create an account
      tags: 
          - developers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
        200:
          description: Error creating this account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
        400:
          description: Request body data faulty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
        500:
          description: Server error saving user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
  /api/auth/login:
    post:
      description: For Users to start a session
      tags: 
          - developers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        200:
          description: User session successfuly started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponseToken'
        400:
          description: Request body data faulty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
...
components:
  schemas:
    UserRegister:
      type: object
      required: 
        - name
        - lastname
        - username
        - password
        - fav_crypto
      properties:
        name:
          type: string
          example: Javier
        lastname:
          type: string
          example: Duarte
        username:
          description: Username is unique, if this user is already taken it won't create the account
          type: string
          example: JavierDuarteC
        password:
          description: Password has to have at least 8 chars in total. At least 1 capital, at least 1 lower case, at least 1 number
          type: string
          example: 1234Abcd
        fav_crypto:
          description: Favorite crypto ID
          type: integer
          example: 12 
    UserLogin:
      type: object
      required: 
        - username
        - password
      properties:
        username:
          description: Username is unique, if this user is already taken it won't create the account
          type: string
          example: JavierDuarteC
        password:
          description: Password has to have at least 8 chars in total. At least 1 capital, at least 1 lower case, at least 1 number
          type: string
          example: 1234Abcd
...
    BasicResponseToken:
      type: object
      required:
        - success
        - message
        - token
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: User logged in
        token:
          type: string
          example: sdasDAifngHLHth-342noefsDFSfd
    BasicResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Your query was executed
...